+++ projects/swimlane/ngx-cron/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## HEAD (unreleased)
 
-## 4.0.0 (2023-01-20)
+## 4.0.0 (2022-01-20)
 
 - Breaking: require new version of `@swimlane/ngx-ui`
 - Feature: show timezone popup when timezone is supplied